backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
agent-framework
python-dotenv>=1.0.0
nest-asyncio>=1.5.0
gunicorn>=21.2.0
